requirements.txt
astroid==2.1.0
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.1.0
django-rest-framework==0.1.0
django-taggit==0.23.0
djangorestframework==3.9.1
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.2
python-decouple==3.1
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
wrapt==1.11.0
